pyproject.toml
[tool.poetry]
name = "shoplist_api"
version = "0.1.0"
description = ""
authors = ["vggog <gluhovkich@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
pydantic = "^1.10.2"
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
python-environ = "^0.4.54"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
